Should be easier to follow how this works now
This should be behaviorally identical to the previous implementation, except way faster, use memory way more efficiently, be drastically easier to read, and also I commented the hell out of it. As an aside, we can probably do something to make sure every language with a config file has at least one key in this without us needing to do anything (would be the language code), if that interests anyone. Inb4 this hasn't actually been used in months