Prefer fs promises over fs sync API
File IO operations are inherantly asynchronous, as JS has to ask the OS to do something, and then there's nothing more for JS to do until the OS is done. When using functions from  the fs sync API, JS will make the system call and then stop and wait and do absolutely nothing until the OS completes the task.This is bad. Instead, we want to use promises, particularly in code run during normal bot runtime e.g. command bodies. When using the promises API, the calls are async, so JS will ask OS to do the file task, but then it will suspend that thread and be able to go handle any other JS stuff that needs to happen, like other command invocations or handling events from anywhere, until the OS completes the task and the thread with the resolved promise is picked back up from the task queue. This is much better, and will especially be important for scaling. I don't think we've seen issues from this yet (none that I'm aware of anyway), but we would've eventually